user access control completed

+++ app/Http/Controllers/PostsController.php
@@ -9,6 +9,16 @@ use App\Post;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except'=>['index','show']]);
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -73,6 +83,10 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post= Post::find($id);
+        //check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized page.');
+        }
         return view('posts.edit')->with('post',$post);
     }
 
@@ -108,6 +122,10 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        //check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized page.');
+        }
         $post->delete();
         return redirect('/posts')->with('success','Post deleted.');
     }
